Add notification mark as read action
